# backend/app/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HUBMI API — Małopolska Social Innovation Hub",
    "description": "Backend aplikacyjny: matchmaking potrzeb społecznych, innowacje, przepływy pracy, głosowanie i piloty.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/pilots": {
      "post": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Utworzenie inicjatywy pilotażowej (tylko administrator)",
        "operationId": "create_pilot_api_v1_pilots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/pilots/{pilot_id}/transition": {
      "post": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Zmiana statusu cyklu życia pilota (tylko administrator)",
        "operationId": "transition_pilot_status_api_v1_pilots__pilot_id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pilot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pilot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pilots/{pilot_id}/volunteer": {
      "post": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Zgłoszenie się jako wolontariusz do projektu",
        "operationId": "register_volunteer_api_v1_pilots__pilot_id__volunteer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pilot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pilot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Rezygnacja z wolontariatu",
        "operationId": "cancel_volunteer_api_v1_pilots__pilot_id__volunteer_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pilot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pilot Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pilots/{pilot_id}/accept-offer": {
      "post": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Akceptacja zaoferowanego miejsca z listy oczekujących",
        "operationId": "accept_place_offer_api_v1_pilots__pilot_id__accept_offer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pilot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pilot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VolunteerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pilots/{pilot_id}/feedback": {
      "post": {
        "tags": [
          "Pilotaże i Wolontariat"
        ],
        "summary": "Ocena satysfakcji beneficjenta lub wolontariusza",
        "operationId": "submit_feedback_api_v1_pilots__pilot_id__feedback_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pilot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pilot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SatisfactionFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SatisfactionFeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ideas/{idea_id}/thread": {
      "get": {
        "tags": [
          "Dyskusje, Powiadomienia i Adaptacje instytucjonalne"
        ],
        "summary": "Wątek dyskusyjny przypisany do pomysłu",
        "operationId": "get_idea_thread_api_v1_ideas__idea_id__thread_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscussionThreadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/threads/{thread_id}/messages": {
      "post": {
        "tags": [
          "Dyskusje, Powiadomienia i Adaptacje instytucjonalne"
        ],
        "summary": "Wysłanie wiadomości w wątku dyskusyjnym",
        "operationId": "post_thread_message_api_v1_threads__thread_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadMessageCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "Dyskusje, Powiadomienia i Adaptacje instytucjonalne"
        ],
        "summary": "In-app powiadomienia dla bieżącego użytkownika",
        "operationId": "get_notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "type": "array",
                  "title": "Response Get Notifications Api V1 Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications/{notif_id}/read": {
      "post": {
        "tags": [
          "Dyskusje, Powiadomienia i Adaptacje instytucjonalne"
        ],
        "summary": "Oznaczenie powiadomienia jako przeczytane",
        "operationId": "mark_notification_read_api_v1_notifications__notif_id__read_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notif_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notif Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adaptations": {
      "post": {
        "tags": [
          "Dyskusje, Powiadomienia i Adaptacje instytucjonalne"
        ],
        "summary": "Formularz adaptacji innowacji przez instytucję",
        "operationId": "create_adaptation_api_v1_adaptations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstitutionAdaptationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstitutionAdaptationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscussionThreadResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ThreadMessageResponse"
            },
            "type": "array",
            "title": "Messages",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "idea_id",
          "title",
          "created_at"
        ],
        "title": "DiscussionThreadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstitutionAdaptationRequest": {
        "properties": {
          "solution_id": {
            "type": "integer",
            "title": "Solution Id"
          },
          "beneficiaries": {
            "type": "string",
            "title": "Beneficiaries"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "resources": {
            "type": "string",
            "title": "Resources"
          },
          "budget": {
            "type": "string",
            "title": "Budget"
          },
          "constraints": {
            "type": "string",
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "solution_id",
          "beneficiaries",
          "location",
          "resources",
          "budget",
          "constraints"
        ],
        "title": "InstitutionAdaptationRequest"
      },
      "InstitutionAdaptationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "solution_id": {
            "type": "integer",
            "title": "Solution Id"
          },
          "beneficiaries": {
            "type": "string",
            "title": "Beneficiaries"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "resources": {
            "type": "string",
            "title": "Resources"
          },
          "budget": {
            "type": "string",
            "title": "Budget"
          },
          "constraints": {
            "type": "string",
            "title": "Constraints"
          },
          "draft_adaptation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Adaptation"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "solution_id",
          "beneficiaries",
          "location",
          "resources",
          "budget",
          "constraints",
          "created_at"
        ],
        "title": "InstitutionAdaptationResponse"
      },
...
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "message",
          "is_read",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
      "PilotCreateRequest": {
        "properties": {
          "solution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution Id"
          },
          "idea_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idea Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "budget_declared": {
            "type": "number",
            "title": "Budget Declared",
            "default": 0.0
          },
          "partners": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partners"
          },
          "test_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Plan"
          },
          "max_volunteers": {
            "type": "integer",
            "title": "Max Volunteers",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "PilotCreateRequest"
      },
      "PilotResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "budget_declared": {
            "type": "number",
            "title": "Budget Declared"
          },
          "budget_approved": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Approved"
          },
          "accountable_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountable Owner"
          },
          "partners": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partners"
          },
          "test_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Plan"
          },
          "max_volunteers": {
            "type": "integer",
            "title": "Max Volunteers"
          },
          "registered_volunteers_count": {
            "type": "integer",
            "title": "Registered Volunteers Count"
          },
          "waiting_list_count": {
            "type": "integer",
            "title": "Waiting List Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "budget_declared",
          "max_volunteers",
          "registered_volunteers_count",
          "waiting_list_count",
          "created_at"
        ],
        "title": "PilotResponse"
      },
      "PilotTransitionRequest": {
        "properties": {
          "target_status": {
            "type": "string",
            "pattern": "^(review|recruitment_funding|pilot|evaluation|dissemination|unavailable)$",
            "title": "Target Status"
          },
          "budget_approved": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Approved"
          },
          "accountable_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accountable Owner"
          }
        },
        "type": "object",
        "required": [
          "target_status"
        ],
        "title": "PilotTransitionRequest"
      },
...
      "SatisfactionFeedbackRequest": {
        "properties": {
          "role": {
            "type": "string",
            "pattern": "^(beneficiary|volunteer)$",
            "title": "Role"
          },
          "rating": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "improvements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvements"
          }
        },
        "type": "object",
        "required": [
          "role",
          "rating"
        ],
        "title": "SatisfactionFeedbackRequest"
      },
      "SatisfactionFeedbackResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pilot_id": {
            "type": "integer",
            "title": "Pilot Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "improvements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Improvements"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pilot_id",
          "user_id",
          "role",
          "rating",
          "created_at"
        ],
        "title": "SatisfactionFeedbackResponse"
      },
...
      "ThreadMessageCreateRequest": {
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ThreadMessageCreateRequest"
      },
      "ThreadMessageResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "thread_id": {
            "type": "integer",
            "title": "Thread Id"
          },
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "thread_id",
          "author_id",
          "author_name",
          "content",
          "created_at"
        ],
        "title": "ThreadMessageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VolunteerResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pilot_id": {
            "type": "integer",
            "title": "Pilot Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "skills_confirmed": {
            "type": "boolean",
            "title": "Skills Confirmed"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "offered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offered At"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pilot_id",
          "user_id",
          "status",
          "skills_confirmed",
          "position",
          "created_at"
        ],
        "title": "VolunteerResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}